Always fetch data.json, don't load it from cache
The default cache mode for the Fetch API in browsers uses the disk cache for 'fresh' content. While data.json is less than a week old it qualifies as 'fresh'.
Since the status will be updated more frequently than that we need to use a different mode.

'no-cache' always makes a request to the server. If there's a matching entry then the request might be a conditional request.
The cache is updated on a 200 response, and the content of the cache used on a 304 response.

References: https://developer.mozilla.org/en-US/docs/Web/API/RequestInit, https://developer.mozilla.org/en-US/docs/Web/HTTP/Guides/Caching#fresh_and_stale_based_on_age