Script to update data.json with the current availability statuses
The serverside script and the client both use the same code to 'normalise' the availability data.
In this context, normalisation means setting the status to 'unavailable' for missing timestamps.

This means the client can infer that services are currently unavailable when there isn't recent enough data in data.json,
and the serverside script can fill in the missing gap using the same logic when it eventually come online again.