Remove the unnecessary 'earliestFindGrainedData' field
The purpose of this was to show that there was data missing from the start of the record which should be filled in.
However, this won't ever actually happen, and we can rely on the earliest fine grained datapoint being the earliest we expect.

Say we start recording at 9am on the 1st of the month for a particular namespace.

We have data from 9am on the 1st; the earliest datapoint actually available is the earliest we expect.

If data isn't recorded from 10am to 11am then `normalise()` will fill in the gap.
There's no ambiguity about when to fill in:
    data is missing after 10am because there's data before then;
    and no data is missing before 9am before there _isn't_ any data before then.

What if data stops being recorded at 11pm on the 1st.
If `normalise()` is called at 1am on the 2nd then it will fill in the 2 hour gap of missing data, as expected.
However, what if `normalise()` is called at 1am on the 3rd. The code to delete old data will delete the record at 11pm on the 1st,
so how can a future invokation of `normalise()` differentiate between this scenario (expect data on the 2nd, and since it's missing fill it in)
and the scenario we had earlier (don't expect data before 9am on the 1st because we hadn't started before then)?

This is what 'earliestFineGrainedData' was supposed to solve; it would record that we didn't expect anything before 9am on the 1st,
and when old data was deleted it would be moved forwards as appropriate.
However it isn't actually needed. Gaps are filled in before old data is deleted, so there would never be ambiguity requiring this field.
That call to `normalise()` at 1am on the 3rd would first fill in the missing data between 11pm on the 1st and now,
and then delete old data. There will never be a gap in the record at the start, only at the end.