Script to show almost all changes - exclude automated one
The automated changes that are made every 5 minutes aren't interesting during development.
This script includes the regex to exclude those commits.

It assumes the git alias `all` is set up. It is on my dev machines, which is good enough for me.
If anyone else wants to use this (for some reason) then they can easily swap `all` out for `log`.